Emed: merge getAnnotationSection() methods
