+++ husky-fhir/husky-fhir-emed-ch/husky-fhir-emed-ch-epr/src/main/java/org/projecthusky/fhir/emed/ch/epr/resource/pml/ChEmedEprCompositionPml.java
@@ -92,22 +92,6 @@ public class ChEmedEprCompositionPml extends ChEmedEprComposition {
     }
 
     /**
-     * Returns the annotation section; if missing, it creates it.
-     *
-     * @return the annotation section.
-     */
-    public SectionComponent getAnnotationSection() {
-        var section = getSectionByLoincCode(ANNOTATION_SECTION_CODE_VALUE);
-        if (section == null) {
-            section = this.addSection();
-            section.getCode().addCoding(new Coding(FhirSystem.LOINC,
-                                                   ANNOTATION_SECTION_CODE_VALUE, "Annotation comment"));
-            section.setTitle("Comment");
-        }
-        return section;
-    }
-
-    /**
      * Returns the list with medication statement, medication request, medication dispense and observation.
      *
      * @return the list with medication statement, medication request, medication dispense and observation
